@@ -5,7 +5,7 @@
   "author": "Serhii <boonya> Buinytskyi",
   "license": "MIT",
   "engines": {
-    "node": "~18.14",
+    "node": "~18.15",
     "npm": "~9.5"
   },
   "scripts": {
@@ -41,13 +41,13 @@
     "@graphql-codegen/cli": "^3.3.0",
     "@graphql-codegen/schema-ast": "^3.0.1",
     "@graphql-eslint/eslint-plugin": "^3.18.0",
-    "@storybook/addon-a11y": "^7.0.2",
-    "@storybook/addon-actions": "^7.0.2",
-    "@storybook/addon-essentials": "^7.0.2",
-    "@storybook/addon-interactions": "^7.0.2",
-    "@storybook/addon-links": "^7.0.2",
+    "@storybook/addon-a11y": "^7.0.4",
+    "@storybook/addon-actions": "^7.0.4",
+    "@storybook/addon-essentials": "^7.0.4",
+    "@storybook/addon-interactions": "^7.0.4",
+    "@storybook/addon-links": "^7.0.4",
     "@storybook/jest": "^0.1.0",
-    "@storybook/react-webpack5": "^7.0.2",
+    "@storybook/react-webpack5": "^7.0.4",
     "@storybook/testing-library": "^0.1.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
@@ -60,7 +60,7 @@
     "css-loader": "^6.7.3",
     "css-mediaquery": "^0.1.2",
     "dotenv": "^16.0.3",
-    "eslint": "^8.37.0",
+    "eslint": "^8.38.0",
     "eslint-config-bluedrop": "^9.0.0",
     "eslint-import-resolver-babel-module": "^5.3.2",
     "eslint-plugin-babel": "^5.3.1",
@@ -83,9 +83,10 @@
     "jest-environment-jsdom": "^29.5.0",
     "msw": "^1.2.1",
     "msw-storybook-addon": "^1.8.0",
-    "storybook": "^7.0.2",
+    "raw-loader": "^4.0.2",
+    "storybook": "^7.0.4",
     "terser-webpack-plugin": "^5.3.7",
-    "webpack": "^5.78.0",
+    "webpack": "^5.79.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.13.2",
     "webpack-merge": "^5.8.0"
