@@ -23,22 +23,13 @@ IssueDetails.decorators = [
 	wrapper(ISSUE_DATA, FetchIssueContextProvider),
 ];
 
-export function AddReactionFail(args) {
+export function MutationFails(args) {
 	return <Component {...args} />;
 }
-AddReactionFail.decorators = [
+MutationFails.decorators = [
 	wrapper(ISSUE_DATA, FetchIssueContextProvider),
 ];
-AddReactionFail.parameters = {msw: {handlers: {
+MutationFails.parameters = {msw: {handlers: {
 	ADD_REACTION_MUTATION: mutation(ADD_REACTION_MUTATION, {errors: [{message: 'Failed to add reaction.'}]}),
-}}};
-
-export function RemoveReactionFail(args) {
-	return <Component {...args} />;
-}
-RemoveReactionFail.decorators = [
-	wrapper(ISSUE_DATA, FetchIssueContextProvider),
-];
-RemoveReactionFail.parameters = {msw: {handlers: {
 	REMOVE_REACTION_MUTATION: mutation(REMOVE_REACTION_MUTATION, {errors: [{message: 'Failed to remove reaction.'}]}),
 }}};
