@@ -4,28 +4,23 @@ import {REACTION_CONTENT, REACTIONS} from '@src/constants';
 export default {
 	component: Component,
 	args: {
-		reactions: [REACTION_CONTENT.hooray, REACTION_CONTENT.rocket],
-		selected: [],
+		selected: [REACTION_CONTENT.hooray, REACTION_CONTENT.rocket],
+		pending: [],
 	},
 	argTypes: {
-		reactions: {
-			control: 'inline-check',
-			options: REACTIONS,
-		},
 		selected: {
 			control: 'inline-check',
 			options: REACTIONS,
 		},
+		pending: {
+			control: 'inline-check',
+			options: REACTIONS,
+		},
+		onAdd: {table: {disable: false}},
+		onRemove: {table: {disable: false}},
 	},
 };
 
 export function List(args) {
 	return <Component {...args} />;
 }
-
-export function Interactive(args) {
-	return <Component {...args} />;
-}
-Interactive.argTypes = {
-	onClick: {table: {disable: true}},
-};
