@@ -1,24 +1,54 @@
-import './index.css';
 import List from './List';
-import Manage from './Manage';
-import PropTypes from 'prop-types';
-import {REACTIONS} from '@src/constants';
+import {useState, useCallback} from 'react';
+import useAddReaction from '@src/providers/AddReaction/useAddReaction';
+import {useFetchIssueContext} from '@src/providers/FetchIssue/ContextProvider';
+import useRemoveReaction from '@src/providers/RemoveReaction/useRemoveReaction';
 
-export default function Reactions({selected, onAdd, onRemove, ...props}) {
-	return (
-		<div className="reactions-container" {...props}>
-			<Manage selected={selected} onAdd={onAdd} onRemove={onRemove} />
-			<List aria-label="Selected reactions" reactions={selected} />
-		</div>
-	);
+export default function Reactions(props) {
+	const [pending, setPending] = useState([]);
+
+	const {data} = useFetchIssueContext();
+	const {id} = data;
+
+	const includePending = useCallback((name) => {
+		setPending((prev) => [...prev, name]);
+	}, []);
+
+	const excludePending = useCallback((name) => {
+		setPending((prev) => prev.filter((i) => i !== name));
+	}, []);
+
+	const [addReaction] = useAddReaction();
+	const [removeReaction] = useRemoveReaction();
+	const reactions = data.reactions.map(({content}) => content);
+
+	const onAdd = useCallback(async (reaction) => {
+		try {
+			includePending(reaction);
+			await addReaction({issueId: id, reaction});
+		}
+		catch (err) {
+			// eslint-disable-next-line no-alert
+			alert(err.message);
+		}
+		finally {
+			excludePending(reaction);
+		}
+	}, [addReaction, excludePending, id, includePending]);
+
+	const onRemove = useCallback(async (reaction) => {
+		try {
+			includePending(reaction);
+			await removeReaction({issueId: id, reaction});
+		}
+		catch (err) {
+			// eslint-disable-next-line no-alert
+			alert(err.message);
+		}
+		finally {
+			excludePending(reaction);
+		}
+	}, [excludePending, id, includePending, removeReaction]);
+
+	return <List selected={reactions} pending={pending} onAdd={onAdd} onRemove={onRemove} {...props} />;
 }
-
-Reactions.propTypes = {
-	onAdd: PropTypes.func.isRequired,
-	onRemove: PropTypes.func.isRequired,
-	selected: PropTypes.arrayOf(PropTypes.oneOf(REACTIONS)),
-};
-
-Reactions.defaultProps = {
-	selected: [],
-};
