@@ -1,3 +1,3 @@
-require('@testing-library/jest-dom');
+import '@testing-library/jest-dom';
 
 window.NODE_ENV = process.env.NODE_ENV;
